requirements.txt
PySimpleGUI
cryptography
pycryptodome
jsonschema
reportlab
